Aff. carte amélioré, Ajout recherche par genre
