+++ src/components/CarteInfoStatic.vue
@@ -1,0 +1,39 @@
+<template>
+    <v-card @click="doSomething()" v-bind:title="titre">
+        <v-card-title v-if="titre.length > 19" class="d-flex overline justify-center">{{titre.substring(0,19)}}...
+        </v-card-title>
+        <v-card-title v-else class="d-flex overline justify-center">{{titre}}
+        </v-card-title>
+        <v-img align-end  v-bind:src="img" width="200" height="300" raised="true"></v-img>
+        <v-card-text>
+        <v-row align="center">
+          <v-rating
+            :value="score/2"
+            color="amber"
+            half-increments
+            dense
+            size="14"
+            readonly
+          ></v-rating>
+          <div class="grey--text ml-4">{{score}}</div>
+        </v-row>
+        </v-card-text>
+    </v-card>
+</template>
+
+<script>
+export default {
+  props: {
+    titre: String,
+    img: String,
+    score: Number
+  },
+  data: () => ({
+  }),
+  methods: {
+    doSomething () {
+      console.log(this.titre)
+    }
+  }
+}
+</script>
