+++ src/main.js
@@ -1,5 +1,6 @@
 import Vue from 'vue'
 import App from './App.vue'
+import 'material-design-icons-iconfont/dist/material-design-icons.css'
 import vuetify from './plugins/vuetify'
 
 Vue.config.productionTip = false
